Only enable formatting if a style has been selected
Added a new option “clang-format+-always-enable” which defaults to nil.
If clang-format+-always-enable is nil, then only enable formatting if
there is a .clang-format (or _clang-format) file in the source tree, or
if clang-format-style is not set to “file”. This makes clang-format+ do
what I believe most people would expect it to do by default.

The mode line indicator will be “cf+” is formatting is enabled,
otherwise “cf-”.

Closes #2.